test: Rename `test!` to `test_ingest_then_query!` and make it reusable
