+++ server/tests/typo_correction.rs
@@ -24,44 +24,6 @@ mod common;
 
 use crate::common::*;
 
-macro_rules! test {
-    ($sentence:tt $([$option:tt])* $(LANG($ingest_lang:expr))?, $examples:tt $(LANG($query_lang:expr))?) => {
-        init_logging();
-        let executor = make_test_executor();
-
-        exec!(executor -> PUSH "messages" "user:1" "chat:1" $sentence $(LANG($ingest_lang))?);
-        exec!(executor -> TRIGGER consolidate);
-
-        $(test!(internal_ $option: executor, $sentence);)*
-
-        for (needle, should_match) in $examples.into_iter() {
-            assert!(!needle.contains("{"), "Needle shouldn’t contain '{{', make sure you formatted the string correctly.");
-
-            let response = exec!(executor -> QUERY "messages" "user:1" needle $(LANG($query_lang))?);
-            if should_match {
-                assert_eq!(response, ["chat:1"], "Did not find {needle:?} in {:?}", $sentence);
-            } else {
-                assert_eq!(response, vec![] as Vec<&str>, "Found {needle:?} in {:?}", $sentence);
-            }
-        }
-
-        // Drop the executor so it’s not debug-printed multiple times on
-        // failure if this macro was invoked multiple times in the same test.
-        // We could scope this code in a code block (`{ … }`) but it’d be an
-        // implementation detail one might remove in the future unknowingly.
-        drop(executor);
-    };
-
-    (internal_ ensure_no_stopword: $executor:tt, $sentence:ident) => {
-        // Sanity check: ensure no stopword was provided (could make
-        // examples pass for the wrong reason).
-        assert_eq!(
-            exec!($executor -> COUNT "messages" "user:1" "chat:1"),
-            Ok($sentence.split_ascii_whitespace().count() as u32)
-        );
-    }
-}
-
 /// This test sentence contains words that are 3–10 characters-long, while
 /// not being stopwords. It’s not realistic but not having stopwords avoids
 /// false positives in tests.
@@ -76,7 +38,7 @@ const ASTRONOMY_WORDS: &str = "sun moon comet nebula pulsars asteroid satellite 
 #[ignore = "Known issue (FIXME)"]
 fn test_search_allows_typos() {
     #[rustfmt::skip]
-    test!(ASTRONOMY_WORDS [ensure_no_stopword], [
+    test_ingest_then_query!(ASTRONOMY_WORDS [ensure_no_stopword], [
         // 3-letter word, distance = 1.
         ("sum", false),
         // 4-letter word, distance = 1.
@@ -105,7 +67,7 @@ fn test_search_allows_typos() {
 #[ignore = "Not supported yet (FIXME)"]
 fn test_search_term_order_insignificant() {
     #[rustfmt::skip]
-    test!(ASTRONOMY_WORDS [ensure_no_stopword], [
+    test_ingest_then_query!(ASTRONOMY_WORDS [ensure_no_stopword], [
         ("satellite pulsars nebula", true),
         (&format!("missing {ASTRONOMY_WORDS}"), true),
     ]);
